pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "deepagents-dash"
version = "0.1.0"
description = "AI Agent Web Interface with Filesystem and Canvas Visualization"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your@email.com"}
]
keywords = ["ai", "agent", "dashboard", "visualization", "deepagents", "dash", "plotly"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: User Interfaces",
]

dependencies = [
    "dash>=2.0.0",
    "dash-mantine-components>=0.14.0",
    "plotly>=5.0.0",
    "pandas>=1.3.0",
    "matplotlib>=3.5.0",
    "Pillow>=9.0.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
deepagents = [
    "deepagents>=0.1.0",
]

dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

all = [
    "deepagents-dash[deepagents,dev]",
]

[project.urls]
Homepage = "https://github.com/yourusername/deepagents-dash"
Documentation = "https://github.com/yourusername/deepagents-dash/blob/main/README.md"
Repository = "https://github.com/yourusername/deepagents-dash"
"Bug Tracker" = "https://github.com/yourusername/deepagents-dash/issues"

[project.scripts]
deepagents-dash = "deepagents_dash.cli:main"

[tool.setuptools]
packages = ["deepagents_dash"]

[tool.setuptools.package-data]
deepagents_dash = [
    "assets/*",
    "templates/*",
    "config_template.py",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
